Update page.js
minor edit on the initial state of the loading variable. It was true which means that on start it will be false and then it will be true then false. Which will cause data state to load before useEffect hook that is why initial state should be true and then it will be false.